+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changes
 =======
 
+
+0.1.4 -- 2016-08-21
+-------------------
+* Better terminal output
+* Fixed issue #5: QRCode.terminal() uses a special output function (if it
+  detects Windows) to support MS Windows which may not support ANSI escape codes.
+
+
 0.1.3 -- 2016-08-20
 -------------------
 * Added command line script "segno"